# docs/api/m2office-apps.openapi.yaml
...
openapi: 3.1.0
info:
  title: M2Office — 外部のアプリの API
  version: 1.3.0
  description: |
    外のシステム（販売管理・M2Medical など）が M2Office を使うための API です（仕様書 第13.4.1節）。

    **使い方**
    1. 会社の管理者が、管理者ページの「接続」の「**外部のアプリ**」でアプリを登録します。**鍵（`m2oa_` で始まる）はそのとき一度だけ出ます。**
    2. 管理者が、アプリに許す**機能**と機能ごとの設定を選び、「この内容で許す」で承認します。承認するまで、どの機能も使えません。
    3. アプリは `Authorization: Bearer <鍵>` を付けて、**画面と同じ `/v1` の道**を、会社の M2Office の名前で呼びます。

    | 機能 | 道 | 状態 |
    |---|---|---|
    | 会社の基本情報を読む（`company.profile`） | `GET /v1/company/profile` | 使える |
    | 商品の一覧を読む（`inventory.catalog`） | `GET /v1/inventory/catalog` | 使える（在庫管理を入れた会社） |
    | 販売を知らせる（`inventory.sales`） | `POST /v1/inventory/sales-events` | 使える（在庫管理を入れた会社） |
    | アカウントを結び付ける（`accounts.link`） | `/v1/accounts/…` | 使える |
    | ナレッジを検索する（`knowledge.search`） | `POST /v1/knowledge/search` | 使える |
    | 問い合わせを受ける（`inquiries.intake`） | `POST /v1/inquiries/intake` | 使える（問い合わせの記録を入れた会社） |
    | 社内のお知らせを出す（`notices.post`） | `POST /v1/notices` ほか | 使える |
    | 入庫を知らせる（`inventory.receipts`） | `POST /v1/inventory/receipts` | 使える（在庫管理を入れた会社） |
    | 予約の空きを読む・予約を入れる（`reservations.book`） | `/v1/reservations/…` | 使える（予約の管理を入れた会社） |
    | 会員のポイントを付ける・使う（`members.points`） | `POST /v1/members/points` | 使える（会員の管理を入れた会社） |
    | 公開したコラムを読む（`columns.read`） | `GET /v1/columns/published` ほか | 使える（コラムの作成を入れた会社） |
    | 社内規程を登録・改定する（`knowledge.rules`） | `PUT /v1/knowledge/rules/{ref}` ほか | 使える |
...

    **共通の決まり**
...
    - 回数はアプリごとに 1 分 120 回までです。超えたら 429 と `Retry-After` を返します。
...
servers:
...
security:
  - appKey: []
...
paths:
...
  /v1/notices:
    post:
      tags: [社内のお知らせ]
      operationId: postNotice
      summary: 社内のお知らせを出す（機能 notices.post）
      description: |
        画面と同じ形のお知らせを、アプリとして出します（出した人は「外部のアプリ（名前）」）。宛先は、管理者が承認した範囲（全員か、決めたグループ）だけで、範囲の外は 403 です。
        Chat には投稿しません（お知らせだけ）。リンクは https だけです。題名は 60 字、本文は 1000 字までです。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, title]
              properties:
                eventId: { type: string, minLength: 1, maxLength: 100 }
                title: { type: string, maxLength: 60 }
                body: { type: string, maxLength: 1000 }
                link: { type: string, format: uri, pattern: '^https://' }
                all: { type: boolean, description: 全員に出す（承認で全員を許したときだけ） }
                groupIds: { type: array, items: { type: string } }
                dueOn: { type: string, format: date, description: 締切 }
                until: { type: string, format: date, description: 載せる最後の日（省けば締切の日、締切も無ければ 14 日） }
            example: { eventId: notice-2026-1010-1, title: 年末調整の書類を出してください, body: 11 月 15 日までに総務へ。, all: true, dueOn: '2026-11-15' }
      responses:
        '201':
          description: 出した（同じ eventId なら前と同じ答え）
          content:
            application/json:
              example: { id: 6b1d2f3e-1c2a-4f5b-9e8d-7a6b5c4d3e2f, until: '2026-11-15' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: 同じ eventId を別の中身で受け付け済み、または処理の途中 }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /v1/reservations:
    post:
      tags: [予約]
      operationId: createReservation
      summary: 予約を入れる（機能 reservations.book）
      description: |
        結び付いた本人（機能 accounts.link）として予約を入れ、本人の Google カレンダーに予定を入れます（`calendar` が結果）。1 件 14 日まで・90 日先までです。重なれば 409 と次の空きを返します。
        本人が予約を使えない（利用範囲の外）ときは 403、承認していない予約できるものは 404 です。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, bindingId, itemId, startAt, endAt]
              properties:
                eventId: { type: string, minLength: 1, maxLength: 100 }
                bindingId: { type: string, pattern: '^[A-Za-z0-9_-]{32}$' }
                itemId: { type: string }
                startAt: { type: string, format: date-time }
                endAt: { type: string, format: date-time }
                purpose: { type: string, maxLength: 100 }
      responses:
        '201':
          description: 予約した（同じ eventId なら前と同じ答え）
          content:
            application/json:
              example: { id: rsv-1, itemId: rsi-1, startAt: '2026-10-12T01:00:00.000Z', endAt: '2026-10-12T02:00:00.000Z', calendar: added }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: その時間は埋まっている（次の空きを返す）。または同じ eventId を別の中身で受け付け済み
          content:
            application/json:
              example: { error: その時間はもう埋まっています, nextFree: { startAt: '2026-10-12T02:00:00.000Z', endAt: '2026-10-12T03:00:00.000Z' } }
        '410': { description: 結び付きが無効（binding_invalid） }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /v1/jobs:
    post:
      tags: [業務の依頼]
      operationId: createJob
      summary: 業務を依頼する（機能 jobs.run）
      description: |
        結び付いた本人（機能 accounts.link）として、承認した業務だけを依頼します。承認した危険度の上限を超える業務・本人の利用範囲の外の業務・無効にした業務は 403 です。
        承認の段に来たら止まります（状態 `awaiting_approval`）。承認は本人の画面だけで行い、この API からは承認できません。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, bindingId, agentId]
              properties:
                eventId: { type: string, minLength: 1, maxLength: 100 }
                bindingId: { type: string, pattern: '^[A-Za-z0-9_-]{32}$' }
                agentId: { type: string }
                input: { type: object, additionalProperties: true, description: 業務の入力（業務ごとの形） }
      responses:
        '202':
          description: 受け付けた（同じ eventId なら前と同じ答え）
          content:
            application/json:
              example: { runId: 8c1e…, status: queued }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: 会社で AI が使えない、または同じ eventId を別の中身で受け付け済み }
        '410': { description: 結び付きが無効（binding_invalid） }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    appKey:
      type: http
      scheme: bearer
      description: |
        外部のアプリを登録したときに一度だけ出る鍵（`m2oa_` で始まる 37 字）。`Authorization: Bearer <鍵>` で送ります。
        M2Office は鍵のハッシュだけを持ちます。なくしたら、会社の管理者が「鍵を出し直す」で新しい鍵を出します（前の鍵はすぐ使えなくなります）。
  responses:
    Forbidden:
      description: このアプリには、この機能を許していない（管理者が「外部のアプリ」で機能を選んで承認する）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'このアプリには許していない機能です' }
    BadRequest:
      description: 形が違う（どの項目かを返す）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'quantity は 1 以上の整数にしてください', field: 'lines[0].quantity' }
    NotFound:
      description: 鍵が違う・アプリを止めた・会社が止まっている（停止・緊急停止・解約）・呼んだ名前の会社と鍵の会社が合わない・（在庫の機能で）在庫管理を切った。どれも同じ答え（理由を分けない）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'not found' }
    TooManyRequests:
      description: アプリごとに 1 分 120 回を超えた（機能ごとに、さらに絞ることがある）
      headers:
        Retry-After:
          description: 次に送ってよいまでの秒数
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        field: { type: string, description: 形が違った項目 }
...